requirements.txt
streamlit==1.30.0
langchain-groq==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0
openai==1.3.7